Move PinchableImage to a new file
+-------+-------+------------------------------------------------------------------------+
| P     | Type  | Task Description                                                       |
+-------+-------+------------------------------------------------------------------------+
| 1     | Item  | Moving PinchableImage and using useWindowDimensions() hook breaks the  |
|       |       | PinchableImage pinch-to-zoom functionality.                            |
+-------+-------+------------------------------------------------------------------------+
| 1.1   | Task  | Move PinchableImage back to the HomeScreen and verify that the root    |
|       |       | cause of the issue is useWindowDimensions()                            |
+-------+-------+------------------------------------------------------------------------+
| 1.2   | Task  | Replace useWindowDimensions() hook with Dimensions.window()            |
+-------+-------+------------------------------------------------------------------------+
